requirements.txt
asgiref==3.2.10
asn1crypto==1.3.0
base58==0.2.5
bit==0.4.3
bitcash==0.5.2
boto3==1.14.41
botocore==1.17.41
cashaddress==1.0.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coincurve==13.0.0
crypto-django==0.1.1
cytoolz==0.10.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-heroku==0.3.1
django-resized==0.3.11
django-storages==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.15.2
eth-hash==0.2.0
eth-typing==1.3.0
eth-utils==1.2.1
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
pycryptodome==3.9.8
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toolz==0.10.0
urllib3==1.25.9
whitenoise==5.2.0
